+++ kcommons/src/main/java/com/kairos/dto/gdpr/data_inventory/AssetTypeOrganizationLevelDTO.java
@@ -13,7 +13,7 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AssetTypeOrganizationLevelDTO {
 
-    private BigInteger id;
+    private Long id;
 
     @NotBlank(message = "error.message.name.notNull.orEmpty")
     @Pattern(message = "error.message.number.and.special.character.notAllowed", regexp = "^[a-zA-Z\\s]+$")
@@ -24,9 +24,9 @@ public class AssetTypeOrganizationLevelDTO {
     private List<AssetTypeOrganizationLevelDTO> subAssetTypes = new ArrayList<>();
 
 
-    public BigInteger getId() { return id; }
+    public Long getId() { return id; }
 
-    public void setId(BigInteger id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
